# docs/mvp-1/http-contract-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B-01 HTTP — кандидат для согласования",
    "version": "mvp1-http-draft.1",
    "description": "Кандидат, не реализация. См. http-contract-draft.md."
  },
  "servers": [
    {
      "url": "https://contract-not-deployed.invalid"
    }
  ],
  "paths": {
    "/mvp1/browser/context": {
      "get": {
        "operationId": "browserContext",
        "summary": "Получить CSRF-контекст",
        "description": "ПРЕДЛОЖЕНИЕ B-01. Не реализовано. Семантика: http-contract-draft.md.",
        "tags": [
          "browser-read"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Получить CSRF-контекст",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "Set-Cookie": {
                "description": "HttpOnly CSRF-контекст, не новая сессия",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserContext"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "invalid_request"
            ]
          },
          "500": {
            "description": "internal_error",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "internal_error"
            ]
          },
          "403": {
            "description": "invalid_origin",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "invalid_origin"
            ]
          },
          "503": {
            "description": "context_capacity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "description": "Секунды до явной попытки, если известны; не команда retry",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "context_capacity"
            ]
          }
        },
        "parameters": [
          {
            "name": "__Host-photoagent",
            "in": "cookie",
            "required": false,
            "description": "Secure HttpOnly-cookie; не доступна JavaScript",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "__Host-photoagent-context",
            "in": "cookie",
            "required": false,
            "description": "Secure HttpOnly-cookie; не доступна JavaScript",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/mvp1/browser/session": {
      "post": {
        "operationId": "browserSession",
        "summary": "Создать сессию или повторить выдачу cookie",
        "description": "ПРЕДЛОЖЕНИЕ B-01. Не реализовано. Семантика: http-contract-draft.md.",
        "tags": [
          "browser"
        ],
        "security": [
          {
            "BrowserContextCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "204": {
            "description": "Создать сессию или повторить выдачу cookie",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "Set-Cookie": {
                "description": "Secure; HttpOnly; SameSite=Lax; Path=/; без Domain",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "session_unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "session_unavailable"
            ]
          },
          "422": {
            "description": "invalid_request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "invalid_request"
            ]
          },
          "500": {
            "description": "internal_error",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "internal_error"
            ]
          },
          "403": {
            "description": "invalid_origin, invalid_csrf, invalid_context",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "invalid_origin",
              "invalid_csrf",
              "invalid_context"
            ]
          },
          "409": {
            "description": "request_id_conflict, creation_retired, operation_capacity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "request_id_conflict",
              "creation_retired",
              "operation_capacity"
            ]
          },
          "503": {
            "description": "session_capacity",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "description": "Секунды до явной попытки, если известны; не команда retry",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "session_capacity"
            ]
          }
        },
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "description": "Точный допустимый HTTPS origin",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "__Host-photoagent",
            "in": "cookie",
            "required": false,
            "description": "Secure HttpOnly-cookie; не доступна JavaScript",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "__Host-photoagent-context",
            "in": "cookie",
            "required": false,
            "description": "Secure HttpOnly-cookie; не доступна JavaScript",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationRequest"
              }
            }
          }
        }
      }
    },
...
    "/mvp1/internal/session": {
      "get": {
        "operationId": "sessionSnapshot",
        "summary": "Восстановить полный снимок",
        "description": "ПРЕДЛОЖЕНИЕ B-01. Не реализовано. Семантика: http-contract-draft.md.",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Восстановить полный снимок",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "session_unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "session_unavailable"
            ]
          },
          "422": {
            "description": "invalid_request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "invalid_request"
            ]
          },
          "500": {
            "description": "internal_error",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "x-error-codes": [
              "internal_error"
            ]
          }
        },
        "parameters": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "OperationRequest": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "description": "ID явного действия; при retry сохраняется",
            "examples": [
              "22222222-2222-4222-8222-222222222222"
            ],
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "OperationRequest",
        "type": "object",
        "examples": [
          {
            "request_id": "22222222-2222-4222-8222-222222222222"
          }
        ]
      },
...
      "Text": {
        "minLength": 1,
        "type": "string"
      },
...
      "BrowserContext": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "mvp1-http-draft.1",
            "description": "Несогласованный HTTP-конверт; не версия предметного Answer",
            "examples": [
              "mvp1-http-draft.1"
            ],
            "title": "Contract Version",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ID только доступной сессии, не bearer",
            "examples": [
              null
            ],
            "title": "Session Id"
          },
          "csrf_token": {
            "$ref": "#/components/schemas/Text",
            "description": "Привязан к HttpOnly-контексту; не токен сессии",
            "examples": [
              "synthetic-csrf"
            ]
          }
        },
        "required": [
          "contract_version",
          "session_id",
          "csrf_token"
        ],
        "title": "BrowserContext",
        "type": "object",
        "examples": [
          {
            "contract_version": "mvp1-http-draft.1",
            "session_id": null,
            "csrf_token": "synthetic-csrf"
          }
        ]
      },
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": "1.1",
            "description": "Версия ошибки",
            "examples": [
              "1.1"
            ],
            "title": "Schema Version",
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/Text",
            "description": "Безопасный код из контракта маршрута",
            "examples": [
              "session_busy"
            ]
          },
          "message": {
            "$ref": "#/components/schemas/Text",
            "description": "Публичное сообщение без секретов",
            "examples": [
              "Запрос обрабатывается"
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Только авторизованная сессия",
            "examples": [
              null
            ],
            "title": "Session Id"
          },
          "request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ID отправки, если доступен",
            "examples": [
              null
            ],
            "title": "Request Id"
          },
          "session_status": {
            "anyOf": [
              {
                "enum": [
                  "ready",
                  "in_progress",
                  "awaiting_clarification",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Состояние доступной сессии",
            "examples": [
              null
            ],
            "title": "Session Status"
          },
          "active_execution_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Доступное активное выполнение",
            "examples": [
              null
            ],
            "title": "Active Execution Id"
          }
        },
        "required": [
          "schema_version",
          "code",
          "message",
          "session_id",
          "request_id",
          "session_status",
          "active_execution_id"
        ],
        "title": "ApiError",
        "type": "object",
        "examples": [
          {
            "schema_version": "1.1",
            "code": "session_unavailable",
            "message": "Действие не выполнено",
            "session_id": null,
            "request_id": null,
            "session_status": null,
            "active_execution_id": null
          },
          {
            "schema_version": "1.1",
            "code": "invalid_csrf",
            "message": "Действие не выполнено",
            "session_id": null,
            "request_id": null,
            "session_status": null,
            "active_execution_id": null
          },
          {
            "schema_version": "1.1",
            "code": "answer_not_found",
            "message": "Действие не выполнено",
            "session_id": null,
            "request_id": null,
            "session_status": null,
            "active_execution_id": null
          }
        ]
      },
...
    },
    "securitySchemes": {
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cookie сессии передаёт сервер UI; не доступен JS"
      },
      "OwnerBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Отдельное право владельца; сессионный bearer не подходит"
      },
      "BrowserContextCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-photoagent-context"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PhotoAgent-CSRF"
      }
    }
  }
}